Compile with latest rust

+++ src/libstegpeg/lib.rs
@@ -8,19 +8,18 @@
                      ***   *           ***/
 
 
+#![crate_name = "stegpeg"]
 #![crate_type = "rlib"]
 #![crate_type = "dylib"]
-#![crate_id = "stegpeg#0.0.0"]
 
-#![feature(macro_rules)]
-#![allow(ctypes)]
-// #![feature(globs)]
+#![feature(libc)]
+#![allow(improper_ctypes)]
 
 // pub use libjpeg::*;
 
 pub mod libjpeg;
 
 #[no_mangle]
-pub extern fn printme(me: uint) {
+pub extern fn printme(me: usize) {
   println!("HI: {}", me);
 }
